Rename AddCourse2Activity to AddCourseActivity
