+++ app/src/main/AndroidManifest.xml
@@ -80,21 +80,6 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".AddCourseActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.GenericTemplate.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".CourseList"
             android:exported="true"
             android:label="@string/app_name"
@@ -155,7 +140,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".AddCourse2Activity"
+            android:name=".AddCourseActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.GenericTemplate.NoActionBar">
