+++ app/src/main/java/com/presenters/AddCoursePresenter.java
@@ -3,5 +3,5 @@ package com.presenters;
 import com.models.course.CourseManagement;
 
 public class AddCoursePresenter {
-//    CourseManagement management
+    CourseManagement management;
 }
